agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "A Python agent for the SOC system"

authors = [
    { name = "mohit kumar", email = "mohitdevx@proton.me" },
]

dependencies = [
    "aiohttp>=3.13.1",
    "pyyaml>=6.0.3",
    "msgpack>=1.1.2"
]



